Recovered from Eclipse.
git-svn-id: https://subversion.assembla.com/svn/cmosher@553 884d43a4-ffbe-47e9-90ca-fd8c37d7eb32

+++ play/src/Player.java
@@ -70,7 +70,7 @@ public class Player implements Immutable
 //		bufrd.close();
 
 		byte[] rb = new byte[0x100];
-		for (byte i = 0; i < rb.length; i++)
+		for (int i = 0; i < rb.length; i++)
         {
             rb[i] = i;
         }
